Cadastro, alteração, exclusão, findById de usuario

+++ src/main/java/com/ElasTechProjetoFinal/repository/TecnicoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TecnicoRepository extends JpaRepository<Tecnico, Integer> {
+public interface TecnicoRepository extends JpaRepository<Tecnico, Long> {
 }
